keep filter state in useState
