+++ packages/explorer/src/app/(explorer)/[chainName]/worlds/[worldAddress]/interact/InteractForm.tsx
@@ -1,55 +1,49 @@
 "use client";
 
-import { useQueryState } from "nuqs";
-import { AbiFunction, AbiItem, Hex, toFunctionHash } from "viem";
-import { useDeferredValue, useEffect, useMemo, useState } from "react";
+import { useSearchParams } from "next/navigation";
+import { AbiFunction, AbiItem, Hex } from "viem";
+import { useDeferredValue, useMemo, useState } from "react";
 import { hexToResource } from "@latticexyz/common";
 import IBaseWorldAbi from "@latticexyz/world/out/IBaseWorld.sol/IBaseWorld.abi.json";
 import { Input } from "../../../../../../components/ui/Input";
-import { Skeleton } from "../../../../../../components/ui/Skeleton";
-import { useHashState } from "../../../../hooks/useHashState";
 import { useSystemAbisQuery } from "../../../../queries/useSystemAbisQuery";
 import { useWorldAbiQuery } from "../../../../queries/useWorldAbiQuery";
-import { SystemContent } from "./content/SystemContent";
-import { FunctionSidebarItem } from "./sidebar/FunctionSidebarItem";
-import { SystemSidebarItem } from "./sidebar/SystemSidebarItem";
+import { FunctionsContent } from "./content/FunctionsContent";
+import { SidebarContent } from "./sidebar/SidebarContent";
 
 function isFunction(abi: AbiItem): abi is AbiFunction {
   return abi.type === "function";
 }
 
-type System = {
+export type System = {
   systemId: string;
   name: string;
   namespace: string;
   functions: AbiFunction[];
 };
 
-type NamespaceSection = {
+export type NamespaceSection = {
   namespace: string;
   systems: System[];
 };
 
-type FilteredFunctions = {
+export type FilteredFunctions = {
   core: System[];
   namespaces: NamespaceSection[];
 };
 
 export function InteractForm() {
-  const [hash] = useHashState();
-  const { data, isFetched } = useWorldAbiQuery();
-  const { data: systemData } = useSystemAbisQuery();
-  const [expandedSystems, setExpandedSystems] = useState<Record<string, boolean>>({
-    root: true,
-  });
-  const [filterValue, setFilterValue] = useQueryState("function", { defaultValue: "" });
+  const searchParams = useSearchParams();
+  const { data: worldAbiData, isFetched: isWorldAbiFetched } = useWorldAbiQuery();
+  const { data: systemAbis, isFetched: isSystemAbisFetched } = useSystemAbisQuery();
+  const isFetched = isWorldAbiFetched && isSystemAbisFetched;
+  const [filterValue, setFilterValue] = useState(searchParams.get("filter") || "");
   const deferredFilterValue = useDeferredValue(filterValue);
 
   const filteredSystemFunctions = useMemo<FilteredFunctions>(() => {
-    if (!systemData) return { core: [], namespaces: [] };
+    if (!systemAbis) return { core: [], namespaces: [] };
 
     const searchLower = deferredFilterValue.toLowerCase();
-
     const coreFunctions = (IBaseWorldAbi as AbiItem[]).filter((item): item is AbiFunction => {
       if (!isFunction(item)) return false;
       const matchesName = `core_core_${item.name}`.toLowerCase().includes(searchLower);
@@ -68,7 +62,7 @@ export function InteractForm() {
           ]
         : [];
 
-    const systemsByNamespace = Object.entries(systemData).reduce<Record<string, System[]>>(
+    const systemsByNamespace = Object.entries(systemAbis).reduce<Record<string, System[]>>(
       (acc, [systemId, systemAbi]) => {
         const { namespace, name: systemName } = hexToResource(systemId as Hex);
         const namespaceName = namespace || "root";
@@ -109,86 +103,7 @@ export function InteractForm() {
       core: coreSection,
       namespaces: namespaceSections,
     };
-  }, [systemData, deferredFilterValue]);
-
-  const toggleSystem = (systemId: string) => {
-    setExpandedSystems((prev) => {
-      const newState = { ...prev };
-      const isExpanding = !prev[systemId];
-
-      if (isExpanding === false) {
-        const namespace = systemId;
-        const systems = filteredSystemFunctions.namespaces.find((ns) => ns.namespace === namespace)?.systems || [];
-        systems.forEach((system) => {
-          newState[system.systemId] = false;
-        });
-      }
-
-      newState[systemId] = isExpanding;
-      return newState;
-    });
-  };
-
-  // Expand matching systems when the filter value is changed
-  useEffect(() => {
-    if (!deferredFilterValue) {
-      setExpandedSystems({ root: true });
-      return;
-    }
-
-    const newExpandedState: Record<string, boolean> = {};
-
-    if (filteredSystemFunctions.core.length > 0) {
-      newExpandedState.core = true;
-    }
-
-    filteredSystemFunctions.namespaces.forEach(({ namespace, systems }) => {
-      if (systems.length > 0) {
-        newExpandedState[namespace] = true;
-        systems.forEach((system) => {
-          newExpandedState[system.systemId] = true;
-        });
-      }
-    });
-
-    setExpandedSystems(newExpandedState);
-  }, [deferredFilterValue, filteredSystemFunctions]);
-
-  // Expand matching systems when opened with a hash
-  useEffect(() => {
-    if (!hash || !systemData) return;
-
-    const coreFunction = (IBaseWorldAbi as AbiItem[]).find((item) => {
-      if (item.type !== "function") return false;
-      return toFunctionHash(item) === hash;
-    });
-
-    if (coreFunction) {
-      setExpandedSystems((prev) => ({
-        ...prev,
-        root: false,
-        core: true,
-      }));
-      return;
-    }
-
-    for (const [systemId, systemAbi] of Object.entries(systemData)) {
-      const functionWithHash = systemAbi.find((item) => {
-        if (item.type !== "function") return false;
-        return toFunctionHash(item) === hash;
-      });
-
-      if (functionWithHash) {
-        const { namespace } = hexToResource(systemId as Hex);
-        setExpandedSystems((prev) => ({
-          ...prev,
-          [namespace || "root"]: true,
-          [systemId]: true,
-        }));
-        break;
-      }
-    }
-  }, [hash, systemData]);
+  }, [systemAbis, deferredFilterValue]);
 
   return (
     <>
@@ -199,112 +114,24 @@ export function InteractForm() {
             <Input
               type="search"
               placeholder="Filter by namespace, system, or function..."
-              value={deferredFilterValue}
+              value={filterValue}
               onChange={(evt) => setFilterValue(evt.target.value)}
             />
           </div>
 
-          <ul className="mt-4 max-h-max overflow-y-auto pb-4 pr-4">
-            {!isFetched &&
-              Array.from({ length: 6 }).map((_, index) => {
-                return (
-                  <li key={index} className="pt-2">
-                    <Skeleton className="h-[30px]" />
-                  </li>
-                );
-              })}
-
-            {filteredSystemFunctions.namespaces.map(({ namespace, systems }: NamespaceSection) => (
-              <SystemSidebarItem
-                key={namespace}
-                name={namespace}
-                isNamespace
-                isExpanded={expandedSystems[namespace]}
-                onToggle={() => toggleSystem(namespace)}
-                functionCount={systems.reduce((total, system) => total + system.functions.length, 0)}
-              >
-                <ul className="ml-4 space-y-1">
-                  {systems.map((system: System) => {
-                    if (system.functions.length === 0) return null;
-
-                    const isExpanded = expandedSystems[system.systemId];
-                    return (
-                      <SystemSidebarItem
-                        key={system.systemId}
-                        name={system.name}
-                        isExpanded={isExpanded}
-                        onToggle={() => toggleSystem(system.systemId)}
-                        functionCount={system.functions.length}
-                      >
-                        <ul className="mt-0 space-y-1">
-                          {system.functions.map((abi: AbiFunction) => (
-                            <FunctionSidebarItem key={toFunctionHash(abi)} abi={abi} hash={hash} />
-                          ))}
-                        </ul>
-                      </SystemSidebarItem>
-                    );
-                  })}
-                </ul>
-              </SystemSidebarItem>
-            ))}
-
-            {filteredSystemFunctions.core.map((system: System) => {
-              if (system.functions.length === 0) return null;
-
-              const isExpanded = expandedSystems[system.systemId];
-              return (
-                <SystemSidebarItem
-                  key={system.systemId}
-                  name={system.name}
-                  isExpanded={isExpanded}
-                  onToggle={() => toggleSystem(system.systemId)}
-                  functionCount={system.functions.length}
-                >
-                  <ul className="mt-0 space-y-1">
-                    {system.functions.map((abi: AbiFunction) => (
-                      <FunctionSidebarItem key={toFunctionHash(abi)} abi={abi} hash={hash} />
-                    ))}
-                  </ul>
-                </SystemSidebarItem>
-              );
-            })}
-          </ul>
+          <SidebarContent
+            filteredFunctions={filteredSystemFunctions}
+            filterValue={deferredFilterValue}
+            isLoading={!isFetched}
+          />
         </div>
 
-        <div className="w-full overflow-y-auto pl-1 pr-1">
-          {!isFetched && (
-            <div className="space-y-4">
-              <Skeleton className="h-[150px]" />
-              <Skeleton className="h-[150px]" />
-              <Skeleton className="h-[150px]" />
-              <Skeleton className="h-[150px]" />
-              <Skeleton className="h-[150px]" />
-            </div>
-          )}
-
-          {data?.abi && (
-            <>
-              {filteredSystemFunctions.namespaces.map(({ namespace, systems }: NamespaceSection) => (
-                <SystemContent
-                  key={namespace}
-                  name={namespace}
-                  systems={systems}
-                  worldAbi={data.abi as AbiItem[]}
-                  isNamespace
-                />
-              ))}
-
-              {filteredSystemFunctions.core.map((system: System) => (
-                <SystemContent
-                  key={system.systemId}
-                  name={system.name}
-                  functions={system.functions}
-                  worldAbi={data.abi as AbiItem[]}
-                />
-              ))}
-            </>
-          )}
-        </div>
+        <FunctionsContent
+          worldAbi={worldAbiData?.abi}
+          filteredFunctions={filteredSystemFunctions}
+          filterValue={deferredFilterValue}
+          isLoading={!isFetched}
+        />
       </div>
     </>
   );
